requirements.txt
fastapi>=0.100.0,<0.112.0
uvicorn[standard]>=0.20.0,<0.30.0
python-dotenv>=1.0.0,<2.0.0
atlassian-python-api==4.0.3
pydantic>=2.0.0,<3.0.0

# Testing
pytest>=7.0.0,<8.0.0 # For running tests
httpx>=0.23.0,<0.28.0 # For making async HTTP requests in tests
pytest-asyncio # Ensure this is active
python-multipart>=0.0.5,<0.1.0 # For FastAPI/Starlette form parsing, resolves deprecation warning
